convert dependent controller functions to di helper class
